update:  move _app and _document into layout.js
