+++ src/app/layout.js
@@ -1,7 +1,45 @@
-function MyApp({ Component, pageProps }) {
-  // 以下の処理は不要
-  const getLayout = Component.getLayout ?? ((page) => page)
-  return getLayout(<Component {...pageProps} />)
-}
+'use client';
 
-export default MyApp
+import 'swiper/swiper-bundle.css';
+// components
+import { Header } from 'src/components/common/header';
+import { Footer } from 'src/components/common/footer';
+// hooks
+import { useHeaderScroll } from 'src/hooks/useHeaderScroll';
+
+import CssBaseline from '@mui/material/CssBaseline';
+import { CacheProvider } from '@emotion/react';
+import { ThemeProvider } from '@mui/material/styles';
+import theme from 'src/theme';
+import createEmotionCache from 'src/createEmotionCache';
+
+const clientSideEmotionCache = createEmotionCache();
+
+import * as React from 'react';
+
+// export const metadata = {
+//   description: 'Generated by create next app',
+//   viewport: 'width=device-width, initial-scale=1',
+//   icons: {
+//     icon: '/favicon.ico',
+//   },
+// };
+
+export default function RootLayout({ emotionCache = clientSideEmotionCache, children }) {
+  const { isHeaderActive } = useHeaderScroll();
+
+  return (
+    <CacheProvider value={emotionCache}>
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <html lang='ja'>
+          <body>
+            <Header isActive={isHeaderActive} />
+            <main>{children}</main>
+            <Footer />
+          </body>
+        </html>
+      </ThemeProvider>
+    </CacheProvider>
+  );
+}
