--- src/pages/_app.jsx
@@ -1,33 +1,0 @@
-// Next.js
-
-// style
-// import "../styles/globals.css";
-
-import CssBaseline from '@mui/material/CssBaseline';
-import { CacheProvider } from '@emotion/react';
-import { ThemeProvider } from '@mui/material/styles';
-import theme from 'src/theme';
-import createEmotionCache from 'src/createEmotionCache';
-
-const clientSideEmotionCache = createEmotionCache();
-
-export const metadata = {
-  description: 'Generated by create next app',
-  viewport: 'width=device-width, initial-scale=1',
-  icons: {
-    icon: '/favicon.ico',
-  },
-};
-
-export default function App({ Component, emotionCache = clientSideEmotionCache, pageProps }) {
-  return (
-    <>
-      <CacheProvider value={emotionCache}>
-        <ThemeProvider theme={theme}>
-          <CssBaseline />
-          <Component {...pageProps} />
-        </ThemeProvider>
-      </CacheProvider>
-    </>
-  );
-}
